fix setting value on screen and firebase /no lag
fix value: the trick was to transform the cardValue function to accept an argument which is the object we want to caluclate the value from , then set the newData object again immediately after with the current value, then passing this new data (newDataWithAdjustedValues) instead of newData. this new value is then sent inside updateCardQuantity which now handle the the removing of the old  array, adding the new array as well as set the state. this way there is no lag when setting. also yes, firebase doesn't have a way to update a nested array, it has to bedeleted then set again