handle filter, sort and order
the Deck component now can be sorted by name, series, and release date, as well as set ascendin/descending order. this is done by having a chained map in the chain of fucntions that return the output  with a switch case that will return the operation that we need in the sort().

filter is now working by having an if inside a filter chained to the main return function