permanency of currency rate
now currency rate will stay between sessions (user logging in and out)
as this is saved directly in firestore in the user info. this is done in
setUserInfo at first, where the initial data is taken from firebase and
if there is no currency set(since we dion't set one at first run as we
want only certain data to be refreshed) then we set standard usd.

then, setConversionRate is now connected to the dropwdown in Profile and
will set the state and data on firebase . then coversionrate on state
(we dont want the conversion rate saved on firebase as it's variable )