add dinamic names and notes
it adds setCurrentDeckInfo to handle the change of name and note. it
does so by accepting user, the data to change and the key of the data to
change as props, then updateDoc on firebase and set state on
currentDeckInfo.

currentcollectioninfo has been refactored to have 2 more components, collectionNome and CollectionNote, each allow the render of the name and notes as well as editing it.