fix bug on deck1
it was noticed that deck1 info were always overwritten as the
information were set on fetch with the user uid and such, causing to
lose the name and the other data. this has been fixed by assing the logic to add a deck in the right way after fetching. if after fetching the decks array lenght is 0, then we add a deck, removing side logic like setting manually the decks array and the info on firebase since we have a function that does that