serve the values from historical prices to chart
the historical values from the backend and the way they are collected
are missing the quantities from the historical data. the function has
been modified to add this information which is stored in the card object
itself. when i create a new historical data object i will recover the
amount its available today (which might be different tomorrow)

the DeckCard component has been modified to fetch data from the backend
and put the full card value on the chart. it will be modified to also multiply its value by the quantity of that card (which is a data i don't have on the database atm but it has been refactored to store it from today)