fix the chart
the chart has been fixed and now it doesn't show any "socket hang up"
issue thanks to the fact that the query function of the pokemontcg api
allows for multiple seatch in one call, bringing down the numbers of
calls to the api for each users from numberOfDecks * cards to numberOfDecks (effettively making one call for each deck instead of one call for each card). this might become an issue in the future nonetheless, when there are too many decks and too many users.