fix chart and sorting
the sorting has been fixed by having the chart being sorted by Date and not by index
(whcih was not supposed to be used in the first place and kept as a placeholder
until more undertanding of react-charts was archived). This way the data is more representative of what is supposed to be on screen and i don't have to create an index to sort it while i have the Date on it. the chart now shows the Date of when the data has been added on the screen, need to be refactored a bit for better graphics