NEW FEATURE: card view on click
this feature allows the user to click on the card image and have it
shown at better resolution in the center of the screen. more informations could be added but it's unclear from the specification what this should do at the moment.

this commit also add an event listener to the search bar which detects
when the enter key is pressed, if the searchbar is not empty then it
will start the search