BUGFIX: current value of cards in chart
there was a mistake with the current visualizaion of the chart, as it
would show the last value with out of date data, as i wasn't updating
the userDeckInfo.value as well as adjustedValues. this has been
temporarely solved by having a custom function calculate the value of
the current card during the fetch of the deck and return the adjusted
object with the values needed. this object is not set on Firebase at
this stage.