BUGFIX: fix dropdown list for filter
the dropdown has beeen fixed by removing onClick and having onChange on
the <select> JSXcomponent which now makes it work normally. the
filtering logic has now been fixed to the opposite of how it was
originally done, as selecting a key will show only the cards WITH THAT
KEY (previouslu it would remove all the cards with that key, doesn't
seems like) a feature since we have a massive amount of keys depending n
the amount of cards. the filtering has also been expanded, by having each key work shown on screen and clickable: when done, the keyword is removed from the filter, making sure that the user can just remove the last filter if is not the one he was lookign for instead of having to press "clear" to remove all