FEATURE: aggregate all similar card under one card, on click opens overlay with all cards
this has been done in order to minimize visual clutter when too many cards of the
same id. the new system will check on the deck component if the map function has
already encountered a card with the same id, if so it will filter it out by not
returnin it and instead saving its id on a list. Later in the chain of operations it
will check the cards against that list, if there is a match it will return the item
with a "type" as argument (group/collection/search) so that the card component can
decide on a switch case what kind of item to return

Card.TSX has been expanded by having the type of cards being generated on different
components instead of conditionally render its part and make it hard to read