FIX: AddCard.tsx, Alert.tsx and firebase function
addcard has been fixed to have all the necessary functions inside,
previously some were passed from the parent unnecessarily. this allows
for the function now to be called from anywhere without having to add
anything else

Alert now has a z-index of 9999 as it was noted that previously this
woudl clash with preset tailwindcss classes, so now the alert is on top
of everythign at all time when called

firebase function has been tested with a stripped down version and found
it executed and works as needed. needs to be investigated why despite
calling the cards it won't still set them one by one