fix: cloud function sets by set
to optimise the fetch time the cards are now stored in subcolecctions by set. this will allow to query the cards by sets instead that quering the entire collection of all the cards, limiting the amount of cards that needs to be checked (which has to be done for each iteration, so it will eventually become too slow). By limiting the amoutn of cards we query, we should get a faster function